### 2021-10-29 #3869 @whedon
[PRE REVIEW]: RECOLO: A Python package for the reconstruction of surface pressure loads from kinematic fields using the virtual fields method
**Submitting author:** @PolymerGuy (<a href="http://orcid.org/0000-0002-9981-4041">Sindre Nordmark Olufsen</a>)
**Repository:** <a href="https://github.com/PolymerGuy/recolo" target ="_blank">https://github.com/PolymerGuy/recolo</a>
**Version:** v1.0.0
