### 2021-10-30 c6b53b3
Fixed affiliation formatting error
