### 2021-10-30 #3869 @PolymerGuy
Hi @kthyng and @Kevin-Mattheus-Moerman ! I have now added the section "Statement of need" and hopefully fixed the bug related to affiliations above. When compiling the paper I didn't get any warnings on this error but I think it's OK now.
