### 2021-11-19 #3869 @Kevin-Mattheus-Moerman
@ahojukka5 @azzeddinetiba @jeevanjyoti4 @ilyasst @diehlpk @yangbai90 @kegiljarhus @corentin-dev would you be interested in reviewing this work for the [Journal of Open Source Software (JOSS)](https://joss.theoj.org):
 "_RECOLO: A Python package for the reconstruction of surface pressure loads from kinematic fields using the virtual fields method_".

The [smooth review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) takes place on GitHub on a dedicated review issues (e.g. here is [an example](https://github.com/openjournals/joss-reviews/issues/3860)) and focuses on:
- [A short paper](https://github.com/openjournals/joss-papers/blob/joss.03869/joss.03869/10.21105.joss.03869.pdf)
- [The software](https://github.com/PolymerGuy/recolo)

Let me know if you have any questions. Note that given the COVID pandemic we can be quite flexible regarding time needed to review. Thanks!
