### 2021-11-19 #3869 @jeevanjyoti4
I am really sorry. But I don't think I have the expertise to properly
examine this work.

I hope to be able to contribute to the review process in the future.

Best regards,
Jeevan

-----------------------------------------------------------------------------
Dr Jeevanjyoti Chakraborty
Assistant Professor
Department of Mechanical Engineering
Indian Institute of Technology (IIT) Kharagpur
Kharagpur - 721302
West Bengal, India
http://www.facweb.iitkgp.ac.in/~jeevanjyoti/
-----------------------------------------------------------------------------

On Fri, 19 Nov 2021 at 15:52, Kevin Mattheus Moerman <
***@***.***> wrote:

> @ahojukka5 <https://github.com/ahojukka5> @azzeddinetiba
> <https://github.com/azzeddinetiba> @jeevanjyoti4
> <https://github.com/jeevanjyoti4> @ilyasst <https://github.com/ilyasst>
> @diehlpk <https://github.com/diehlpk> @yangbai90
> <https://github.com/yangbai90> @kegiljarhus
> <https://github.com/kegiljarhus> @corentin-dev
> <https://github.com/corentin-dev> would you be interested in reviewing
> this work for the Journal of Open Source Software (JOSS)
> <https://joss.theoj.org>:
> "*RECOLO: A Python package for the reconstruction of surface pressure
> loads from kinematic fields using the virtual fields method*".
>
> The smooth review process
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html> takes
> place on GitHub on a dedicated review issues (e.g. here is an example
> <https://github.com/openjournals/joss-reviews/issues/3860>) and focuses
> on:
>
>    - A short paper
>    <https://github.com/openjournals/joss-papers/blob/joss.03869/joss.03869/10.21105.joss.03869.pdf>
>    - The software <https://github.com/PolymerGuy/recolo>
>
> Let me know if you have any questions. Note that given the COVID pandemic
> we can be quite flexible regarding time needed to review. Thanks!
>
