### 2021-11-23 #3869 @kegiljarhus
Dear Kevin,

Sorry for the late reply on this one. Thank you for considering me as a
reviewer, but this falls outside my field of expertise so I hope someone
else more suitable can contribute.

Regards,
Knut Erik Giljarhus

On Tue, 23 Nov 2021 at 16:30, Kevin Mattheus Moerman <
***@***.***> wrote:

> @jomorlier <https://github.com/jomorlier> @ahojukka5
> <https://github.com/ahojukka5> @azzeddinetiba
> <https://github.com/azzeddinetiba> @ilyasst <https://github.com/ilyasst>
> @diehlpk <https://github.com/diehlpk> @yangbai90
> <https://github.com/yangbai90> @kegiljarhus
> <https://github.com/kegiljarhus> @corentin-dev
> <https://github.com/corentin-dev> would you be interested in reviewing
> this work for the Journal of Open Source Software (JOSS)
> <https://joss.theoj.org>:
> "*RECOLO: A Python package for the reconstruction of surface pressure
> loads from kinematic fields using the virtual fields method*".
>
> The smooth review process
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html> takes
> place on GitHub on a dedicated review issues (e.g. here is an example
> <https://github.com/openjournals/joss-reviews/issues/3860>) and focuses
> on:
>
>    - A short paper
>    <https://github.com/openjournals/joss-papers/blob/joss.03869/joss.03869/10.21105.joss.03869.pdf>
>    - The software <https://github.com/PolymerGuy/recolo>
>
> Let me know if you have any questions. Note that given the COVID pandemic
> we can be quite flexible regarding time needed to review. Thanks!
>
