### 2021-12-07 #3980 @whedon
[REVIEW]: RECOLO: A Python package for the reconstruction of surface pressure loads from kinematic fields using the virtual fields method
**Submitting author:** @PolymerGuy (<a href="http://orcid.org/0000-0002-9981-4041">Sindre Nordmark Olufsen</a>)
**Repository:** https://github.com/PolymerGuy/recolo
**Branch with paper.md** (empty if default branch): JOSS-review
**Version:** v1.0.1
