### 2021-12-07 #3980 @nicoguaro
@PolymerGuy, I have done a first pass on the submission and have the following comments:

- There is no Statement of need on the README or the docs.
- There is a section title for an example without the example.
- There is a broken hyperlink to a reference in the README. Is this one the paper where the approach is described or is it the 2019 one?
- I think that the contribution guidelines could be more specific.

Once I know which reference to check for the approach I will continue the review.
