### 2021-12-08 #3980 @PolymerGuy
@nicoguaro, thanks for the comments. I will implement the comments as soon as possible. Should I push the corrections directly to the master branch or should we use a separate revision branch for the review?

The reference in the README is a placeholder for a paper which is about to be submitted now.
For a description of the approach we use, please see the 2019 paper (https://link.springer.com/article/10.1007/s11340-019-00530-2)
