### 2021-12-08 #3980 @Kevin-Mattheus-Moerman
> @nicoguaro, thanks for the comments. I will implement the comments as soon as possible. Should I push the corrections directly to the master branch or should we use a separate revision branch for the review?

You can use any branch you like this is fully up to you. As long as we compile the paper from the correct branch, e.g. `@whedon generate pdf from branch branchName`
