### 2021-12-08 26f7c97
Revised according to JOSS review comments
