### 2021-12-08 #3980 @PolymerGuy
@nicoguaro I have now made a new branch called JOSS-review for all changes done during the review process. Improvements based on the comments above are found in commit: 26f7c97f4efc2ce93a402c5ac8b5479523883845 .
