### 2022-01-10 #3980 @nicoguaro
> @corentin-dev and @nicoguaro, please let me know if there is anything I can assist you with during the review process.

I'll resume activities tomorrow. So, I guess that I'll resume the review as well this week.
