### 2022-01-11 #3980 @nicoguaro
@PolymerGuy, the repo (or docs) is missing a Statement of Need. Furthermore, I think that the one provided in the paper answers the question of why you wrote the software by not how other people might want to use it.
