### 2022-01-12 #3980 @PolymerGuy
> @PolymerGuy, the repo (or docs) is missing a Statement of Need. Furthermore, I think that the one provided in the paper answers the question of why you wrote the software by not how other people might want to use it.

@nicoguaro  I will address this as soon as possible. Just to make sure everything is clear, the commit: [26f7c97f4efc2ce93a402c5ac8b5479523883845](https://github.com/PolymerGuy/recolo/commit/26f7c97f4efc2ce93a402c5ac8b5479523883845) did not address this in a satisfactory manner?
