### 2022-01-12 #3980 @nicoguaro
> @nicoguaro I will address this as soon as possible. Just to make sure everything is clear, the commit: [26f7c97f4efc2ce93a402c5ac8b5479523883845](https://github.com/PolymerGuy/recolo/commit/26f7c97f4efc2ce93a402c5ac8b5479523883845) did not address this in a satisfactory manner?

Oh, I forgot that we have another branch for the review, sorry about that. The description is definitely much appropriate for a wider audience. The docs and the paper might benefit from being more explicit (and less technical?).
