### 2022-01-28 #3980 @corentin-dev
Hi,

Sorry for the delay, concerning the software, I'm just starting to try to use it.

Please consider rebuilding the paper, no sure if it was updated by whedon.

### Installation

Most of it runs smoothly but,

- pytest does not run (at least on my machine). What I did : `python -m venv env`, `source ./env/bin/activate`, `pip install -r requirements.txt`, `pytest recolo`. I get an error: `E   ModuleNotFoundError: No module named 'natsort'`
- current tests get 37 passed, 9 warnings in 27.44s, lots of `invalid value encountered in true_divide`, can you comment please?
- maybe suggest running `python -m pytest recolo` instead?
- example `python ./examples/Experiments/impact_hammer_experiment.py` needs `SciencePlots` for some style in `matplotlib`
- a minor filename error in `bandwith_vs_grid_pitch.py`
- all examples seem to run

### Comments

The code seems to be well commented, and an automatic tool for coverage is used.

### Documentation

The documentation is automaticaly generated. The documentation is very short and should be enhanced with the way the code can be used. Most of the documentation is made of few examples. But nothing detail the structure of the code, and how to setup other cases.

### Tests

I have no idea how much percentage of the capacities are tested. Tests are present, run smoothly, with some warnings.

#### Examples

Some examples are present, even in the documentation. It can help get a hand of the software.

### Paper

> Other VFM toolkits such as PeriPyVFM are readily available but are focused on different applications.

What different application?

Can you specify the range of use of this tool? To what purpose it can be used?
Last, why `Recolo`?

Best regards,
Corentin Lothode
