### 2022-01-31 4208ee4
Avoiding global pytest when running tests
