+++ docs/install.rst
@@ -39,4 +39,5 @@ These tests are integration and unit tests
 
 If you cloned the repo, you have to call pytest from within the folder::
 
-    pytest recolo
+    python -m pytest recolo
+
