### 2022-01-31 58bca9a
Avoiding divide by zero error. We now have deformation also in the first frame
