### 2022-01-31 3835ea1
Added SciencePlots for nice figures
