### 2022-01-31 b224364
Added code coverage when running tests
