+++ docs/conf.py
@@ -16,9 +16,10 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
-# import os
-# import sys
-# sys.path.insert(0, os.path.abspath('.'))
+import os
+import sys
+sys.path.insert(0, os.path.abspath('.'))
+sys.path.insert(0, os.path.abspath('..'))
 
 
 # -- General configuration ------------------------------------------------
@@ -31,8 +32,11 @@
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = ['sphinx.ext.autodoc',
+'sphinx.ext.autosummary',
     'sphinx.ext.mathjax',
               'sphinxcontrib.bibtex']
+
+autosummary_generate = True  # Turn on sphinx.ext.autosummary
 
 bibtex_bibfiles = ['refs.bib']
 bibtex_reference_style = 'author_year'
