### 2022-01-31 1dded5d
Fixed erroneous documentation string stub
