+++ docs/impactHammer.rst
@@ -1,3 +1,5 @@
+.. _hammer-label:
+
 Experiment with impact hammer
 =============================
 
