+++ docs/impactHammer.rst
@@ -1,5 +1,3 @@
-.. _hammer-label:
-
 Experiment with impact hammer
 =============================
 
