### 2022-02-01 #3980 @PolymerGuy
@Kevin-Mattheus-Moerman @corentin-dev @nicoguaro  We have now implemented all comments except those regarding the documentation and the paper. We are currently working on those changes and they will be available by the end of this week.
The following changes have been made:
* The natsort-error when running pytest appears to be caused by a call to the global pytest installation. This is fixed in: https://github.com/PolymerGuy/recolo/commit/4208ee482b220a7c7c94e036733b7e24f31de101
* The warning "invalid value encountered in true_divide" is now fixed in: https://github.com/PolymerGuy/recolo/commit/f9623e026869fca1fe1be8a61e37c9deab1b96a4
* Divide by zero error was due to a test starting at zero deformation, now fixed in: https://github.com/PolymerGuy/recolo/commit/58bca9a212df2eac22958e1362b62fb3ff8f77a0
* python -m pytest recolo is now the suggested way to run the tests, fixed in:https://github.com/PolymerGuy/recolo/commit/ 4208ee482b220a7c7c94e036733b7e24f31de101
* SciencePlots is now added to requirements.txt in: https://github.com/PolymerGuy/recolo/commit/3835ea12f7f57d1ec0c6cfdf2d233736399ac9e2
* Fixed typo in filename: https://github.com/PolymerGuy/recolo/commit/4f16d39b90ddfa7427464947d746e722f68ab4e0
* I have added coverage to the suggested command for running the tests: https://github.com/PolymerGuy/recolo/commit/b22436490d0e54c78776870efd454551abe9cf4e

Regarding the documentation and the paper, we will add automatic API documentation which parses the docstrings and also provides an overview of the formatting of the package. Further, we are working on a separate section acting as an overview, meant to give the user an understanding of the overall design of this tool as well as suggestions on where to start and relevant examples. The paper will be updated according to the comments from @corentin-dev and @nicoguaro above.
