+++ docs/overview.rst
@@ -1,19 +1,25 @@
 Overview
 ========
-**Recolo** provides tools for reconstructing the pressure field acting on a plate based on
-the deformation of the plate. Currently, this tool is limited to small elastic deformations, but will be extended to
+**RECOLO** provides tools for reconstructing the pressure field acting on a plate based on
+the corresponding deformation measurements. The tools allow the user to perform virtual experiments on synthetically generated data as well as performing pressure reconstruction on experimental datasets of full-field deformation measurements (See figure below).
+
+.. image:: ./figures/recolo.png
+   :scale: 100 %
+   :alt: Illustration of RECOLO
+   :align: center
+
+Currently, this tool is limited to small elastic deformations, but will be extended to
 large plastic deformations in the future. In order to assess the capabilities of this toolkit, a suite of tools
-for performing virtual experiments are provided. The tools allows the user to generate synthetic input for the load
+for performing virtual experiments are provided. The tools allow the user to generate synthetic input for the load
 reconstruction based on the results from finite element simulations, providing strict control over all aspects of the imposed loading
-(spatial and temporal variations) as well as experimental aspects such as sensor noise. The synthetic data is also
-integral for testing of the toolkit, facilitating adoption and modification.
+(spatial and temporal variations) as well as experimental aspects such as signal-to-noise ratio. The synthetic data is also a powerful tool for testing of the toolkit, facilitating adoption, development and validation.
 
 The toolkit can be seen as two complementary parts with the following functionality:
 
 * Tools for determining the dynamic pressure field acting on a thin plate based on deformation measurements
     * In-plane deformation measurements using "the grid method"
     * Slope and deflection measurements using deflectometry
-    * Noise removal using a variety of filters
+    * Controlling the signal-to-noise ratio using a variety of filters
     * Pressure reconstruction using the virtual fields method
 * Tools for generating synthetic input based on dynamic finite element simulations of plates subjected to arbitrary pressure fields.
     * Parser for the finite element suite Abaqus
