### 2022-02-02 #3980 @PolymerGuy
@Kevin-Mattheus-Moerman @corentin-dev @nicoguaro We have now updated the documentation and the paper according to the comments above. For convenience, I have merged the JOSS-review branch with the master branch so that  readthedocs builds the updated documentation. The updated documentation can be found here: https://recolo.readthedocs.io/en/latest/?badge=latest
