### 2022-03-02 #3980 @Kevin-Mattheus-Moerman
@PolymerGuy since the reviewers have recommended acceptance we will now proceed to process this submission for acceptance.

I have proofread your paper and it looks fine. However I do have one recommendation below:
- [ ] I recommend you rephrase `Still, it is of utmost importance to provide structural engineers with a detailed knowledge of loads, understand the  underlying physics and predict how structures respond during extreme loading events.` to something like `Still, it is of the utmost importance to provide structural engineers with a detailed knowledge of loads, an understanding of the  underlying physics, and the ability to predict how structures respond during extreme loading events.`

Further to the above, please work on the following remaining steps:
- [ ] We recommend you proofread the paper yourself once more and in particular verify that the author list is accurate and complete and that the acknowledgements are in order.
- [ ] If you have made any changes please update the paper by calling `@editorialbot generate pdf from branch JOSS-review` (note the recent renaming of our editorial bot).
- [ ] Please confirm what the version tag for this publication should be e.g. is it still `v1.0.0`?
- [ ] Once the above is complete please archive a copy of the software on [ZENODO](https://zenodo.org/). You can do this manually but some find [these automated steps](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) useful. Once you have archived the software please report back here with the ZENODO doi.
- [ ] For the archived version please ensure that the authors name set and order is the same and that the title of the archive also matches that of the paper. Note that you may need to manually edit this information. You may also consider adding ORCID information for authors.

Let me know when the above is finished.
