### 2022-03-03 #3980 @PolymerGuy
@Kevin-Mattheus-Moerman The paper has not been compiled after the review comments have been implemented and approved by the reviewers. All changes made during this review has now been merged to -master.

- I have now proofread the paper and confirmed the list of authors and acknowledgements.
- I tried to compile the paper using editorialbot, but the accepted commands seems to be changed.
- I have now released the software with a tag v1.0.0.
- The software has now been archived using Zenodo, using a tag v1.0.1.
- The Zenodo DOI is: https://doi.org/10.5281/zenodo.6324870
- For the list of authors, I did not find any list of authors for the archive on Zenodo?
