### 2022-03-03 #3980 @Kevin-Mattheus-Moerman
@PolymerGuy apologies, the editorial bot recently changed as you can see. It should work now as shown.

1. Can you clarify some things. Did you say that the changes in the paper have not been reviewed yet? If so I'll ask the reviewers to have a quick look.
2. On the ZENODO archive, you need to edit the title and add your other authors (like on the paper). Have you been able to edit the archive? ZENODO lets you do this.
