### 2022-03-04 b2b1470
Added Citation.cff
This file was added to comply with the guidelines when publishing in Journal of Open Source Software
