+++ CITATION.cff
@@ -1,0 +1,18 @@
+cff-version: 1.1.0
+message: "If you use this software, please cite it as below."
+authors:
+  - family-names: Sindre Nordmark
+    given-names: Olufsen
+    orcid: https://orcid.org/0000-0002-9981-4041
+  - family-names: Rene
+    given-names: Kaufmann
+    orcid: https://orcid.org/0000-0001-6248-7027
+  - family-names: Egil
+    given-names: Fagerholt
+    orcid: https://orcid.org/0000-0002-8051-8381
+  - family-names: Vegard
+    given-names: Aune
+    orcid: https://orcid.org/0000-0003-0905-5131
+title: RECOLO: A Python package for the reconstruction of surface pressure loads from kinematic fields using the virtual fields method
+version: v1.0.1
+date-released: 2022-03-03
