### 2022-03-04 #3980 @PolymerGuy
@Kevin-Mattheus-Moerman The state of the papers is as it was when the reviewers replied that the amendments made were answering their comments.
I have now been able to edit the ZENODO archive and the meta-data now matches the JOSS-paper.
