### 2022-03-04 #3980 @Kevin-Mattheus-Moerman
@openjournals/dev are the archive links no longer presented as URLs (in both the return comment and the DOI listed at the top of this issue)? I think it would be better to have them be URLs to so that I can click on them to verify the link points to the right thing.
