### 2022-03-04 #3980 @xuanxu
The bib file contains references with multiple/duplicated ISBNs.
[This PR](https://github.com/PolymerGuy/recolo/pull/1) should fix it.
