### 2022-03-04 #3980 @xuanxu
> Thanks, @xuanxu did you also see my other query on the urls for DOIs? :point_up:

Saw it now. Yeah we'll add links, at least in the reply comment.
