### 2022-03-05 #3980 @PolymerGuy
@Kevin-Mattheus-Moerman I have now verified the changes made in the PR and ensured that the ISBN numbers are correct. The PR has been accepted and merged.
