### 2022-03-07 #3980 @Kevin-Mattheus-Moerman
@PolymerGuy congratulations on getting this work published in JOSS! Thanks for your patience during this longer than usual review process.

@nicoguaro, @corentin-dev thanks for your review efforts!!!
